+++ src/main/kotlin/icu/windea/pls/config/CwtTagType.kt
@@ -3,7 +3,7 @@ package icu.windea.pls.config
 enum class CwtTagType(
     val id: String
 ) {
-    Predefined("predefined"), // ## tag
+    Predefined("tag"), // ## tag
     TypeKeyPrefix("type key prefix"), // type_key_prefix = xxx
     ;
 }