+++ src/main/kotlin/icu/windea/pls/lang/index/ParadoxDefinitionTypeIndex.kt
@@ -13,8 +13,8 @@ class ParadoxDefinitionTypeIndex : StringStubIndexExtension<ParadoxScriptDefinit
         val INSTANCE by lazy { findStubIndex<ParadoxDefinitionTypeIndex>() }
         val KEY = StubIndexKey.createIndexKey<String, ParadoxScriptDefinitionElement>("paradox.definition.type.index")
 
-        private const val VERSION = 58 //1.3.27
-        private const val CACHE_SIZE = 1024 //180+ in stellaris@3.6
+        private const val VERSION = 59 //1.3.32
+        private const val CACHE_SIZE = 2 * 1024 //180+ in stellaris@3.6
     }
 
     override fun getKey() = KEY