+++ src/main/kotlin/icu/windea/pls/script/structureView/ParadoxScriptTreeElement.kt
@@ -3,16 +3,13 @@ package icu.windea.pls.script.structureView
 import com.intellij.ide.structureView.*
 import com.intellij.ide.structureView.impl.common.*
 import com.intellij.psi.*
+import icu.windea.pls.script.psi.*
 
 abstract class ParadoxScriptTreeElement<T : PsiElement>(element: T) : PsiTreeElementBase<T>(element) {
     protected fun postHandleMemberChildren(children: MutableCollection<StructureViewTreeElement>) {
-        if (children.size <= 1) return
+        if (children.isEmpty()) return
 
-        val hasProperty = children.any { it is ParadoxScriptPropertyTreeElement }
-        val hasValue = children.any { it is ParadoxScriptValueTreeElement }
-        if (hasProperty && hasValue) {
-            //不要在结构视图中显示特殊标签
-            children.removeIf { it is ParadoxScriptValueTreeElement && it.isTag() }
-        }
+        //不要在结构视图中显示特殊标签
+        children.removeIf { it is ParadoxScriptValueTreeElement && it.element?.tagType() != null }
     }
 }